TASK-4: Use FlashBag as requested to confirm task addition / deletion Update `redirectTodoWithMessage` to use session FlashBag too Use Bootstrap modal to nicely display confirmation messages on frontend

+++ src/controllers.php
@@ -11,27 +11,21 @@ $app['twig'] = $app->share($app->extend('twig', function($twig, $app) {
 
 
 /**
- * Retrieve and display the TODO list
- * If there is an issue in parameters, pass it to twig to gracefully display it on frontend and help user
+ * Redirect to the TODO page and set a session message to display
  *
  * @param  Array    $app          Silex $app
  * @param  String   $APIMessage   [Optional] If set, pass an error message to display on frontend
  * @param  String   $messageType  [Optional] One of the bootstrap alert type. Default 'danger'
  * @return The twig template
  */
-function displayTodoListPage($app, $APIMessage = null, $messageType = 'danger') {
-  if (null === $user = $app['session']->get('user')) {
-      return $app->redirect('/login');
-  }
+function redirectTodoWithMessage($app, $APIMessage = null, $messageType = 'danger') {
 
-  $sql = "SELECT * FROM todos WHERE user_id = '${user['id']}'";
-  $todos = $app['db']->fetchAll($sql);
+  // Set message and alert type
+  $app['session']->getFlashBag()->set('message', $APIMessage);
+  $app['session']->getFlashBag()->set('type', $messageType);
 
-  return $app['twig']->render('todos.html', [
-      'todos' => $todos,
-      'message' => $APIMessage,
-      'type' => $messageType
-  ]);
+  // redirect to the TODO page
+  return $app->redirect('/todo');
 }
 
 
@@ -80,7 +74,12 @@ $app->get('/todo/{id}', function ($id) use ($app) {
             'todo' => $todo,
         ]);
     } else {
-        return displayTodoListPage($app);
+        $sql = "SELECT * FROM todos WHERE user_id = '${user['id']}'";
+        $todos = $app['db']->fetchAll($sql);
+
+        return $app['twig']->render('todos.html', [
+            'todos' => $todos
+        ]);
     }
 })
 ->value('id', null);
@@ -105,26 +104,37 @@ $app->get('/todo/{id}/json', function ($id) use ($app) {
 
 $app->post('/todo/add', function (Request $request) use ($app) {
 
-
     if (null === $user = $app['session']->get('user')) {
         return $app->redirect('/login');
     }
-
     $user_id = $user['id'];
     $description = $request->get('description');
+    $confirmed = $request->get('confirmed');
 
     // If no description provided, set a usefull error message and redirect to the todo page
     if (empty($description)) {
-      return displayTodoListPage($app, 'Please add a description.');
+      return redirectTodoWithMessage($app, 'Please add a description.');
     }
 
-    $sql = "INSERT INTO todos (user_id, description) VALUES ('$user_id', '$description')";
-    $app['db']->executeUpdate($sql);
+    // If we have the user confirmation, add the new task in DB
+    if (!empty($confirmed)) {
+      $sql = "INSERT INTO todos (user_id, description) VALUES ('$user_id', '$description')";
+      $app['db']->executeUpdate($sql);
 
-    return $app->redirect('/todo');
+      return $app->redirect('/todo');
+    }
+    // Else set flash bag for confirmation modal
+    else {
+      $app['session']->getFlashBag()->set('add_confirmation', $description);
+      return $app->redirect('/todo');
+    }
+
 });
 
 
+/**
+ * Retrieve an order and toggle its done state
+ */
 $app->post('/todo/toggleState/{id}', function ($id) use ($app) {
     if (null === $user = $app['session']->get('user')) {
         return $app->redirect('/login');
@@ -134,14 +144,15 @@ $app->post('/todo/toggleState/{id}', function ($id) use ($app) {
         $sql = "SELECT * FROM todos WHERE id = '$id'";
         $todo = $app['db']->fetchAssoc($sql);
 
-        // Toggle the done state of the todo
+        // If the task wasn't done, set it done now !
         if (is_null($todo['done_date'])) {
           $updated = $app['db']->executeUpdate("UPDATE `todos` SET `done_date` = NOW() WHERE `todos`.`id` = $id");
-          return ($updated === 1) ? displayTodoListPage($app, 'Good work !', 'success') : displayTodoListPage($app, 'Cannot update your task :/');
+          return ($updated === 1) ? redirectTodoWithMessage($app, 'Good work !', 'success') : redirectTodoWithMessage($app, 'Cannot update your task :/');
         }
+        // ... else remove the done_date to see it pending in the todo list
         else {
           $updated = $app['db']->executeUpdate("UPDATE `todos` SET `done_date` = NULL WHERE `todos`.`id` = $id");
-          return ($updated === 1) ? displayTodoListPage($app, 'Keep on working !', 'info') : displayTodoListPage($app, 'Cannot update your task :/');
+          return ($updated === 1) ? redirectTodoWithMessage($app, 'Keep on working !', 'info') : redirectTodoWithMessage($app, 'Cannot update your task :/');
         }
     } else {
         return $app->redirect('/todo');
@@ -150,10 +161,25 @@ $app->post('/todo/toggleState/{id}', function ($id) use ($app) {
 ->value('id', null);
 
 
-$app->match('/todo/delete/{id}', function ($id) use ($app) {
+$app->post('/todo/delete/{id}', function (Request $request) use ($app) {
 
+  if (null === $user = $app['session']->get('user')) {
+      return $app->redirect('/login');
+  }
+  $user_id = $user['id'];
+  $id = $request->get('id');
+  $confirmed = $request->get('confirmed');
+
+  // If we have the user confirmation, add the new task in DB
+  if (!empty($confirmed)) {
     $sql = "DELETE FROM todos WHERE id = '$id'";
     $app['db']->executeUpdate($sql);
 
+    return redirectTodoWithMessage($app, 'Task removed!', 'info');
+  }
+  // Else set flash bag for confirmation modal
+  else {
+    $app['session']->getFlashBag()->set('delete_confirmation', $id);
     return $app->redirect('/todo');
+  }
 });
